+++ HISTORY.rst
@@ -2,6 +2,12 @@
 History
 =======
 
+0.1.1 (2020-08-10)
+------------------
+
+* Add classifiers: LinearSVC, HistGradientBoostingClassifier, SGDClassifier, RidgeClassifierCV.
+* Modify Readme.md file.
+
 0.1.0 (2020-08-10)
 ------------------